Fix tests
pytest.approx syntax was wrong (left over from removing numpy isclose).
Length units should match the input (cm not m).
Move fixure to conftest.py so that it can be shared.
