+++ tests/test_plasma_source.py
@@ -7,31 +7,6 @@ from random import random
 
 from parametric_plasma_source import PlasmaSource
 
-plasma_params = {
-    "elongation": 1.557,
-    "ion_density_origin": 1.09e20,
-    "ion_density_peaking_factor": 1,
-    "ion_density_pedistal": 1.09e20,
-    "ion_density_seperatrix": 3e19,
-    "ion_temperature_origin": 45.9,
-    "ion_temperature_peaking_factor": 8.06,
-    "ion_temperature_pedistal": 6.09,
-    "ion_temperature_seperatrix": 0.1,
-    "major_radius": 906.0,
-    "minor_radius": 292.258,
-    "pedistal_radius": 0.8 * 292.258,
-    "plasma_id": 1,
-    "shafranov_shift": 44.789,
-    "triangularity": 0.270,
-    "ion_temperature_beta": 6,
-}
-
-
-@pytest.fixture(scope="session")
-def plasma_source():
-    """Make a plasma source to use as a test fixture."""
-    return PlasmaSource(**plasma_params)
-
 
 class TestPlasmaSource:
     """A class to run tests against the plasma source."""
@@ -40,59 +15,57 @@ class TestPlasmaSource:
         """Test the ion density at the magnetic origin."""
         ion_density = plasma_source.ion_density(0.0)
 
-        assert pytest.approx(ion_density, 1.09e20)
+        assert ion_density == pytest.approx(1.09e20)
 
     def test_ion_density_inside_pedestal(self, plasma_source):
         """Test the ion density inside the pedestal."""
-        ion_density = plasma_source.ion_density(0.2)
+        ion_density = plasma_source.ion_density(20.0)
 
-        assert pytest.approx(ion_density, 1.09e20)
+        assert ion_density == pytest.approx(1.09e20)
 
     def test_ion_density_outside_pedestal(self, plasma_source):
         """Test the ion density outside the pedestal."""
-        ion_density = plasma_source.ion_density(2.4)
+        ion_density = plasma_source.ion_density(240.0)
 
-        assert pytest.approx(ion_density, 1.00628584e20)
+        assert ion_density == pytest.approx(1.00629067e20)
 
     def test_ion_density_boundary(self, plasma_source):
         """Test the ion density at the boundary."""
-        boundary = plasma_params["minor_radius"] / 100.0
-        ion_density = plasma_source.ion_density(boundary)
+        ion_density = plasma_source.ion_density(plasma_source.minor_radius)
 
-        assert pytest.approx(ion_density, plasma_params["ion_density_seperatrix"])
+        assert ion_density == pytest.approx(plasma_source.ion_density_seperatrix)
 
     def test_ion_temperature_magnetic_origin(self, plasma_source):
         """Test the ion temperature at the magnetic origin."""
         ion_temperature = plasma_source.ion_temperature(0.0)
 
-        assert pytest.approx(ion_temperature, plasma_params["ion_temperature_origin"])
+        assert ion_temperature == pytest.approx(plasma_source.ion_temperature_origin)
 
     def test_ion_temperature_inside_pedestal(self, plasma_source):
         """Test the ion temperature inside the pedestal."""
-        ion_temperature = plasma_source.ion_temperature(0.2)
+        ion_temperature = plasma_source.ion_temperature(20.0)
 
-        assert pytest.approx(ion_temperature, 45.89987429)
+        assert ion_temperature == pytest.approx(45.89987429)
 
     def test_ion_temperature_outside_pedestal(self, plasma_source):
         """Test the ion temperature outside the pedestal."""
-        ion_temperature = plasma_source.ion_temperature(2.4)
+        ion_temperature = plasma_source.ion_temperature(240.0)
 
-        assert pytest.approx(ion_temperature, 5.45525594)
+        assert ion_temperature == pytest.approx(5.45529258)
 
     def test_ion_temperature_boundary(self, plasma_source):
         """Test the ion temperature at the boundary."""
-        boundary = plasma_params["minor_radius"] / 100.0
-        ion_temperature = plasma_source.ion_temperature(boundary)
+        ion_temperature = plasma_source.ion_temperature(plasma_source.minor_radius)
 
-        assert pytest.approx(
-            ion_temperature, plasma_params["ion_temperature_seperatrix"]
+        assert ion_temperature == pytest.approx(
+            plasma_source.ion_temperature_seperatrix
         )
 
     def test_dt_cross_section(self, plasma_source):
         """Test the dt cross section at a specific temperature."""
         dt_cross_section = plasma_source.dt_xs(4.25e7)
 
-        assert pytest.approx(dt_cross_section, 0.0)
+        assert dt_cross_section == pytest.approx(0.0)
 
     def test_source_to_from_string(self, plasma_source):
         """Test the source can be converted to and from a string."""
